% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conquerDE.R
\name{conquerDE}
\alias{conquerDE}
\title{conquerDE}
\usage{
conquerDE(
  L,
  method = c("edgeRQLF", "edgeRLRT", "DESeq2", "glmGamPoi", "limma", "limmatrend"),
  alpha = 0.05
)
}
\description{
run DE analysis
}
