% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{conquerList-class}
\alias{conquerList-class}
\alias{conquerClass}
\title{conquerList}
\description{
create a S4 class
}
\section{Slots}{

\describe{
\item{\code{counts}}{A counts matrix.}

\item{\code{meta}}{A table with sample meta data (colData), coerced to data.frame.}

\item{\code{contrast}}{A matrix specifying contrast}

\item{\code{model_matrix}}{A matrix specifying experimental design}
}}

