% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_edgeRLRTdeconv.R
\name{run_edgeRLRTdeconv}
\alias{run_edgeRLRTdeconv}
\title{run_edgeRLRTdeconv}
\usage{
run_edgeRLRTdeconv(L, alpha = 0.5)
}
\description{
run edgeR LRT analysis using deconvolution normalization \insertRef{Lun2016}{conquerDE}
}
