% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_edgeRzinbwave.R
\name{run_edgeRzinbwave}
\alias{run_edgeRzinbwave}
\title{apply_edgeRzinbwave}
\usage{
run_edgeRzinbwave(L, alpha = 0.05)
}
\description{
run edgeR with zinbwave weights and LRT for DE (aka zinger method)
}
\details{
see the zinger paper \insertRef{VandenBerge2018}{conquerDE}
}
